+++ src/comp0188_cw2/training/validate_single_epoch.py
@@ -49,8 +49,8 @@ class ValidateSingleEpoch:
             will be the same keys required by the criterion. 
         """
 
-        losses = torch.tensor(0)
-        denom = torch.tensor(0)
+        losses = torch.tensor(0.0, dtype=torch.float32)
+        denom = torch.tensor(0.0, dtype=torch.float32)
         if gpu:
             _device = "cuda"
         else:
